stiled cart item, cart footer and header responsive

+++ backend/app/Http/Controllers/Admin/UsersController.php
@@ -48,7 +48,9 @@ class UsersController extends Controller
      */
     public function update(Request $request, User $user)
     {
-        $user->roles()->sync($request->roles);
+        if (Gate::allows('admin')) {
+            $user->roles()->sync($request->roles);      
+        }
         return redirect(route('admin.users.index'));
     }
 
@@ -60,8 +62,12 @@ class UsersController extends Controller
      */
     public function destroy(User $user)
     {
-        $user->roles()->detach();
-        $user->delete();
+
+        if (Gate::allows('admin')) {
+            $user->roles()->detach();
+            $user->delete();        
+        }
+       
         return redirect( route( 'admin.users.index' ));
     }
 }
